Use KiSystemCall64 as a fallback for kpcr_find2
This fixes Windows initialization on a Windows host.

While introspecting a host (with Bareflank Boxy) if the host CPU has
mitigations in hardware for meltdown, Windows won't use KVA Shadow
(i.e. KPTI) and the `KiSystemCall64Shadow` method to find kpcr will
fail. As a fallback method, we need to also try for `KiSystemCall64`.