Bareflank: Add self domid support for hypercalls
hcall_v2p is currently used to translate a v2p on our domain (i.e.
libvmi's domain), not on the target. Having a `self` domid allows us to
reuse hcalls without having to know our domid. Boxy defines SELF in
`bfsdk/include/bfhypercall.h` and we redefine it here as `BF_DOMID_SELF`.

Note:

We should probably add bareflank's bfsdk as a subproject that cmake
downloads for us and that the bareflank driver links with. This would
allow us to not have to redefine everything and possibly unify the ABI
for hypercalls.