bfsdk: Simplify copying bfhypercall.h
bfsdk should be modified upstream in Boxy or MicroV and copied
over. For now we copy it manually but the build system could handle this
instead, which would also allow us to not have to track files in
bfsdk/* .

Defining `MV_EXCLUDE_DEPRECATED` allows to not have to include the
deprecated API which would also conflict with LibVMI defines.